+++ api/all/build.gradle.kts
@@ -1,6 +1,7 @@
 plugins {
   id("otel.java-conventions")
   id("otel.publish-conventions")
+  id("java-test-fixtures")
 
   id("otel.jmh-conventions")
   id("otel.animalsniffer-conventions")
@@ -17,6 +18,10 @@ dependencies {
 
   testImplementation("edu.berkeley.cs.jqf:jqf-fuzz")
   testImplementation("com.google.guava:guava-testlib")
+  testFixturesApi(project(":testing-internal"))
+  testFixturesApi("junit:junit")
+  testFixturesApi("org.assertj:assertj-core")
+  testFixturesApi("org.mockito:mockito-core")
 }
 
 tasks.test {
