+++ api/all/src/test/java/io/opentelemetry/api/logs/DefaultLoggerTest.java
@@ -5,34 +5,15 @@
 
 package io.opentelemetry.api.logs;
 
-import static org.assertj.core.api.Assertions.assertThatCode;
+class DefaultLoggerTest extends AbstractDefaultLoggerTest {
 
-import io.opentelemetry.api.common.AttributeKey;
-import io.opentelemetry.api.common.Attributes;
-import io.opentelemetry.context.Context;
-import java.time.Instant;
-import java.util.concurrent.TimeUnit;
-import org.junit.jupiter.api.Test;
+  @Override
+  protected LoggerProvider getLoggerProvider() {
+    return DefaultLoggerProvider.getInstance();
+  }
 
-class DefaultLoggerTest {
-
-  @Test
-  void buildAndEmit() {
-    assertThatCode(
-            () ->
-                DefaultLogger.getInstance()
-                    .logRecordBuilder()
-                    .setTimestamp(100, TimeUnit.SECONDS)
-                    .setTimestamp(Instant.now())
-                    .setObservedTimestamp(100, TimeUnit.SECONDS)
-                    .setObservedTimestamp(Instant.now())
-                    .setContext(Context.root())
-                    .setSeverity(Severity.DEBUG)
-                    .setSeverityText("debug")
-                    .setBody("body")
-                    .setAttribute(AttributeKey.stringKey("key1"), "value1")
-                    .setAllAttributes(Attributes.builder().put("key2", "value2").build())
-                    .emit())
-        .doesNotThrowAnyException();
+  @Override
+  protected Logger getLogger() {
+    return DefaultLogger.getInstance();
   }
 }
