+++ api/incubator/src/main/java/io/opentelemetry/api/incubator/logs/ExtendedDefaultLogger.java
@@ -20,7 +20,7 @@ class ExtendedDefaultLogger implements ExtendedLogger {
 
   private ExtendedDefaultLogger() {}
 
-  static Logger getInstance() {
+  static Logger getNoop() {
     return INSTANCE;
   }
